+++ js_bundles/Globals.js
@@ -1043,7 +1043,7 @@ function initialiseLayout() {
 	resizeContainer();
     
     //Set version number:
-    document.getElementById("Version").innerHTML = "FreshWDL - Version: 1.1.6 Alpha. yerren@renerica.com";
+    document.getElementById("Version").innerHTML = "FreshWDL - Version: 1.1.7 Alpha. yerren@renerica.com";
 }
 
 //Set global Graph options